ThirdYear/2025-csc1049-smurphy-dcuparkingavailability/code/parkingBackend/requirements.txt
amqp==5.3.1
asgiref==3.7.2
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
bcrypt==3.2.2
billiard==4.2.1
blinker==1.7.0
celery==5.4.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloud-init==24.4
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.5
cryptography==41.0.7
dbus-python==1.3.2
distro==1.9.0
distro-info==1.7+build1
Django==4.2.11
django-celery-beat==2.7.0
django-cors-headers==4.3.1
django-filter==23.5
django-guardian==2.4.0
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.4.0
ephem==4.1.5
gyp==0.1
httplib2==0.20.4
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itypes==1.2.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
kombu==5.4.2
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysqlclient==2.2.7
netifaces==0.11.0
oauthlib==3.2.2
prompt_toolkit==3.0.50
psycopg2==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycurl==7.45.3
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-memcached==1.58
pytz==2024.1
PyYAML==6.0.1
redis==5.2.1
requests==2.31.0
rich==13.7.1
service-identity==24.1.0
setuptools==68.1.2
six==1.16.0
sqlparse==0.4.4
systemd-python==235
Twisted==24.3.0
typing_extensions==4.10.0
tzdata==2025.1
ubuntu-pro-client==8001
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.13
wheel==0.42.0
zope.interface==6.1
